tests: move waiting for child proc. into exec_pike function
